+++ extensions/wrapper/wrapper-common-mappers/src/test/java/de/sovity/edc/ext/wrapper/api/common/mappers/AssetMapperTest.java
@@ -73,10 +73,10 @@ class AssetMapperTest {
         assertThat(uiAsset.getDataSubcategory()).isEqualTo("General Information About Planning Of Routes");
         assertThat(uiAsset.getDataModel()).isEqualTo("my-data-model-001");
         assertThat(uiAsset.getGeoReferenceMethod()).isEqualTo("my-geo-reference-method");
-        assertThat(uiAsset.getTransportMode()).isEqualTo("my-geo-reference-method");
+        assertThat(uiAsset.getTransportMode()).isEqualTo("my-transport-mode");
         assertThat(uiAsset.getSovereignLegalName()).isEqualTo("my-sovereign");
         assertThat(uiAsset.getGeoLocation()).isEqualTo("my-geolocation");
-        assertThat(uiAsset.getNutsLocation()).isEqualTo(Arrays.asList("my-nuts-location1", "my-nuts-location2"));
+        assertThat(uiAsset.getNutsLocations()).isEqualTo(Arrays.asList("my-nuts-location1", "my-nuts-location2"));
         assertThat(uiAsset.getDataSampleUrls()).isEqualTo(Arrays.asList("my-data-sample-urls1", "my-data-sample-urls2"));
         assertThat(uiAsset.getReferenceFileUrls()).isEqualTo(Arrays.asList("my-reference-files1", "my-reference-files2"));
         assertThat(uiAsset.getReferenceFilesDescription()).isEqualTo("my-additional-description");