+++ utils/catalog-parser/src/main/java/de/sovity/edc/utils/catalog/mapper/DspDataOfferBuilder.java
@@ -48,19 +48,15 @@ public class DspDataOfferBuilder {
                 .map(this::buildContractOffer)
                 .toList();
 
-        var distributions = JsonLdUtils.listOfObjects(dataset, Prop.Dcat.DISTRIBUTION);
-
         var assetProperties = Json.createObjectBuilder(dataset)
                 .remove(Prop.TYPE)
                 .remove(Prop.Odrl.HAS_POLICY)
-                .remove(Prop.Dcat.DISTRIBUTION)
                 .build();
 
 
         return new DspDataOffer(
                 assetProperties,
-                contractOffers,
-                distributions
+                contractOffers
         );
     }