+++ utils/json-and-jsonld-utils/src/main/java/de/sovity/edc/utils/jsonld/vocab/Prop.java
@@ -69,11 +69,15 @@ public class Prop {
         public final String CTX_WRONG_BUT_USED_BY_CORE_EDC = "https://www.w3.org/ns/dcat/";
 
         public final String DATASET = CTX_WRONG_BUT_USED_BY_CORE_EDC + "dataset";
-        public final String DISTRIBUTION = CTX_WRONG_BUT_USED_BY_CORE_EDC + "distribution";
+        public final String DISTRIBUTION = CTX + "distribution";
+        public final String DISTRIBUTION_AS_USED_BY_CORE_EDC = CTX_WRONG_BUT_USED_BY_CORE_EDC + "distribution";
         public final String VERSION = CTX + "version";
         public final String KEYWORDS = CTX + "keyword";
         public final String LANDING_PAGE = CTX + "landingPage";
         public final String MEDIATYPE = CTX + "mediaType";
+        public final String START_DATE = CTX + "startDate";
+        public final String END_DATE = CTX + "endDate";
+        public final String DOWNLOAD_URL = CTX + "downloadURL";
     }
 
     /**
@@ -98,6 +102,12 @@ public class Prop {
         public final String CREATOR = CTX + "creator";
         public final String PUBLISHER = CTX + "publisher";
         public final String LICENSE = CTX + "license";
+        public final String TEMPORAL = CTX + "temporal";
+        public final String ACCRUAL_PERIODICITY = CTX + "accrualPeriodicity";
+        public final String SPATIAL = CTX + "spatial";
+        public final String RIGHTS_HOLDER = CTX + "rightsHolder";
+        public final String RIGHTS = CTX + "rights";
+        public final String RIGHTS_STATEMENT = CTX + "RightsStatement";
     }
 
     /**
@@ -130,34 +140,57 @@ public class Prop {
     }
 
     /**
-     * MDS Vocabulary
+     * Namespace mobilitydcatap as specified in
+     * <a href="https://mobilitydcat-ap.github.io/mobilityDCAT-AP/releases/index.html#namespaces">mobilityDCAT-AP</a>
      */
     @UtilityClass
-    public class Mds {
-        public final String CTX = "http://w3id.org/mds#";
-        public final String DATA_CATEGORY = CTX + "dataCategory";
-        public final String DATA_SUBCATEGORY = CTX + "dataSubcategory";
-        public final String DATA_MODEL = CTX + "dataModel";
-        public final String GEO_REFERENCE_METHOD = CTX + "geoReferenceMethod";
+    public class MobilityDcatAp {
+        public final String CTX = "https://w3id.org/mobilitydcat-ap/";
+        public final String MOBILITY_THEME = CTX + "mobilityTheme";
+
+        @UtilityClass
+        public class DataCategoryProps {
+            public final String CTX = "https://w3id.org/mobilitydcat-ap/mobility-theme/";
+            public final String DATA_CATEGORY = CTX + "data-content-category";
+            public final String DATA_SUBCATEGORY = CTX + "data-content-sub-category";
+        }
+
         public final String TRANSPORT_MODE = CTX + "transportMode";
+        public final String GEO_REFERENCE_METHOD = CTX + "georeferencingMethod";
+        public final String DATA_MODEL = CTX + "mobilityDataStandard";
+
+        // Optional property of mobilitydcatap:mobilityDataStandard
+        public final String SCHEMA = CTX + "schema";
     }
 
     /**
-     * Temporary Context for all MDS specific fields awaiting proper classification within the mobilityDCAT-AP specification
+     * Namespace skos as specified in
+     * <a href="https://mobilitydcat-ap.github.io/mobilityDCAT-AP/releases/index.html#namespaces">mobilityDCAT-AP</a>
      */
     @UtilityClass
-    public class MdsDcatExt {
-        public final String CTX = "https://semantic.sovity.io/mds-dcat-ext#";
+    public class Skos {
+        public final String CTX = "http://www.w3.org/2004/02/skos/core#";
+        public final String PREF_LABEL = CTX + "prefLabel";
+    }
 
-        public final String SOVEREIGN = CTX + "sovereign";
-        public final String GEO_LOCATION = CTX + "geolocation";
-        public final String NUTS_LOCATION = CTX + "nuts-location";
-        public final String DATA_SAMPLE_URLS = CTX + "data-sample-urls";
-        public final String REFERENCE_FILES = CTX + "reference-files";
-        public final String ADDITIONAL_DESCRIPTION = CTX + "additional-description";
-        public final String CONDITIONS_FOR_USE = CTX + "conditions-for-use";
-        public final String DATA_UPDATE_FREQUENCY = CTX + "data-update-frequency";
-        public final String TEMPORAL_COVERAGE_FROM = CTX + "temporal-coverage-from";
-        public final String TEMPORAL_COVERAGE_TO = CTX + "temporal-coverage-to";
+    /**
+     * Namespace adms as specified in
+     * <a href="https://mobilitydcat-ap.github.io/mobilityDCAT-AP/releases/index.html#namespaces">mobilityDCAT-AP</a>
+     */
+    @UtilityClass
+    public class Adms {
+        public final String CTX = "http://www.w3.org/ns/adms#";
+        public final String SAMPLE = CTX + "sample";
+    }
+
+    /**
+     * Namespace rdfs as specified in
+     * <a href="https://mobilitydcat-ap.github.io/mobilityDCAT-AP/releases/index.html#namespaces">mobilityDCAT-AP</a>
+     */
+    @UtilityClass
+    public class Rdfs {
+        public final String CTX = "http://www.w3.org/2000/01/rdf-schema#";
+        public final String LITERAL = CTX + "Literal";
+        public final String LABEL = CTX + "label";
     }
 }