+++ CHANGELOG.md
@@ -16,6 +16,8 @@ please see [changelog_updates.md](docs/dev/changelog_updates.md).
 #### Minor Changes
 - Logginghouse-Client: Add logging-house-client extension 0.2.10
 
+- Migrated MDS fields to mobilityDCAT-AP
+
 #### Patch Changes
 
 ### Deployment Migration Notes